DevSecOpsKB-observability/requirements.txt
python-dotenv
openai
langchain
llama-index
pypdf
PyCryptodome
gradio
graphsignal
